requirements.txt
slack-sdk
slack-bolt
python-dotenv
requests
pymupdf
python-pptx
python-docx
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-openai==0.2.1
langchain-postgres
langchain-text-splitters
psycopg[binary]
pandas
openpyxl